Update: add address to user relationship and crud endpoints

+++ API/src/main/java/com/spartanHardware/service/impl/AddressServiceImpl.java
@@ -24,51 +24,59 @@ public class AddressServiceImpl implements IAddressService {
 
     @Override
     @Transactional
-    public AddressResponseDTO registerAddress(AddressRequestDTO dto) {
+    public AddressResponseDTO registerAddress(AddressRequestDTO dto,User loggedUser) {
         Address address = mapper.toAddress(dto);
-        Address savedAddress = repository.save(address);
+        loggedUser.getAddresses().add(address);
+        address.setUser(loggedUser);
+        address.setDefaultAddress(true);
 
-        return mapper.toDto(savedAddress);
+        return mapper.toDto(repository.save(address));
     }
 
     @Override
     @Transactional
-    public AddressResponseDTO updateAddress(AddressRequestDTO dto, Long id) {
+    public AddressResponseDTO updateAddress(AddressRequestDTO dto, Long id, User
+                                             loggedUser) {
         Address address = getAddressById(id);
-        User user = new User();
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if(!auth.getName().equalsIgnoreCase(user.getUsername()))
-            System.out.println("Usuario no coincide");
+        if(!address.getUser().equals(loggedUser))
+            // TODO: 22/2/2023 Cambiar mensaje por properties
+            throw new RuntimeException("Esta dirección no coincide con el usuario logeado");
+
         Address updatedAddress = mapper.toUpdatedAddress(dto,address);
 
         return mapper.toDto(repository.save(updatedAddress));
     }
 
     @Override
-    public List<AddressResponseDTO> getAllAddresses() {
-        return mapper.toDtoList(repository.findAll());
+    public List<AddressResponseDTO> getAllAddresses(User loggedUser) {
+
+        return mapper.toDtoList(loggedUser.getAddresses());
     }
 
     @Override
     public Address getAddressById(Long id) {
-        if(!repository.findById(id).isPresent())
-            System.out.println("La dirección no coincide");
+        if(repository.findById(id).isEmpty())
+            // TODO: 22/2/2023 Cambiar mensaje por properties
+            throw new RuntimeException("La dirección no coincide");
         return repository.findById(id).get();
     }
 
     @Override
-    public AddressResponseDTO getAddressDtoById(Long id) {
-        return mapper.toDto(getAddressById(id));
+    public AddressResponseDTO getAddressDtoById(Long id, User loggedUser) {
+        Address address = getAddressById(id);
+        if(!address.getUser().equals(loggedUser))
+            // TODO: 22/2/2023 Cambiar mensaje por properties
+            throw new RuntimeException("Esta dirección no coincide con el usuario logeado");
+        return mapper.toDto(address);
     }
 
     @Override
     @Transactional
-    public void deleteAddressById(Long id) {
+    public void deleteAddressById(Long id, User loggedUser) {
         Address address = getAddressById(id);
-        User user = new User();
-        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
-        if(!auth.getName().equalsIgnoreCase(user.getUsername()))
-            System.out.println("Usuario no coincide");
+        if(!address.getUser().equals(loggedUser))
+            // TODO: 22/2/2023 Cambiar mensaje por properties
+            throw new RuntimeException("Esta dirección no coincide con el usuario logeado");
         repository.deleteById(id);
     }
 }
